renamed some files, fixed board so that it can assign X and Os in turn on Board
