+++ debug.log
@@ -1,0 +1,1 @@
+[1229/074923.159:ERROR:crash_report_database_win.cc(428)] unexpected header
