+++ src/App.js
@@ -1,12 +1,12 @@
 import React from "react";
 import logo from "./logo.svg";
 import "./App.css";
-import Hello from "./Hello";
+import Board from "./Board";
 
 function App() {
   return (
     <div className="App">
-      <Hello name="some title" age="23" />
+      <Board name="some title" age="23" />
     </div>
   );
 }
